backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
cryptography>=42.0.0
openai>=1.0.0
